requirements.txt
absl-py==0.14.1
astunparse==1.6.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==2.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.21.0  # Updated to a compatible version for Python 3.10.x
opencv-python==4.5.3.56  # Updated to a compatible version
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
Pillow==8.3.1  # Updated to a compatible version
protobuf==4.25.3
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.26.0  # Updated to a compatible version
rich==13.7.1
six==1.16.0
tensorboard==2.6.0  # Removed tensorflow-related packages
tensorflow==2.6.0  # Removed tensorflow-intel and tensorflow-io-gcs-filesystem
termcolor==1.1.0  # Updated to a compatible version
typing-extensions==4.0.1  # Updated to a compatible version
urllib3==1.26.6  # Updated to a compatible version
Werkzeug==2.0.2  # Updated to a compatible version
wrapt==1.16.0
setuptools>=69.2.0
six>=1.16.0
